Update README.md
I created a dynamic and engaging home page for Puma Brand shoes using React. This sleek design showcases a variety of features, including easy contact options, informative details about Puma, a selection of categories, and a quick 'Shop Now' button for a seamless shopping experience. It's a user-friendly platform that brings the world of Puma shoes to your fingertips